Feat: Improve MarkdownComponentsHydration to accept a dictionary of custom transformers.
